+++ src/utils.h
@@ -23,7 +23,7 @@ MatrixFragment getProcessSparseFragment(Context& ctx, int processId);
 
 DenseMatrix gatherDenseMatrix(Context& ctx, DenseMatrix& matrix, int gatherTo);
 
-int gatherCountGE(Context& ctx, DenseMatrix& matrix, int geValue, int gatherTo);
+int gatherCountGE(Context& ctx, DenseMatrix& matrix, double geValue, int gatherTo);
 
 void verifyPreconditions(int p, int c, Algorithm algorithm);
 };  // namespace utils
